update generic_assay fetching methods
