+++ src/shared/components/oncoprint/ResultsViewOncoprintUtils.tsx
@@ -7,7 +7,7 @@ import {AlterationTypeConstants} from "../../../pages/resultsView/ResultsViewPag
 import {Group} from "../../api/ComparisonGroupClient";
 import * as React from "react";
 import { ISelectOption } from "./controls/OncoprintControls";
-import { Treatment } from "shared/lib/GenericAssayUtils";
+import { Treatment } from "shared/lib/GenericAssayUtils/TreatmentUtils";
 
 export const alterationTypeToProfiledForText:{[alterationType:string]:string} = {
     "MUTATION_EXTENDED": "mutations",
