+++ src/shared/lib/GenericAssayUtils/GenericAssayCommonUtils.ts
@@ -1,0 +1,1 @@
+export const NOT_APPLICABLE_VALUE = "NA";
